test: move macro providers to test registries
